DNM Test if bookworm mkfs works with older grub
It is possible for mkfs.ext4 to create an ext4 filesystem that older
grub cannot understand [0]. There is potential that upgrading the
nodepool-builder image to bookworm will update mkfs.ext4 resulting in
dib being unable to run grub-install within chroots. This could happen
if bookworm's mkfs.ext4 is setting newer options that the grub within
image build chroots doesn't understand.

Test if this is the case.

[0] https://bugs.launchpad.net/ubuntu/+source/grub2/+bug/1844012

Change-Id: I8ad23f4722227523c768c883fe87edb9bb312280