AWS driver: Check for volume existence before calculating quota
The quota calculation of the aws driver occasionally raises NoneType
errors because the volume passed was not found for the VolumeId attached
to the instance. This is because of an outdated lru_cache, where volumes
of new instances aren't available until the caches TTL expires.
In this case, skip the quota calculation for the volume, print a warning
with the instance information, and clear the caches to have a correct
quota calculation in the next run.

Change-Id: Ib58dad787483a1a2e873216b25f4d4aa2abbb47c