AWS driver: Check for volume existence before calculating quota
The quota calculation of the aws driver occasionally raises NoneType
errors because the volume passed was not found for the VolumeId attached
to the instance. This can be because of an outdated lru_cache, where
volumes of new instances aren't available until the caches TTL expires.
In this case, skip the quota calculation for the volume and print a
warning with the instance information.

Change-Id: Ib58dad787483a1a2e873216b25f4d4aa2abbb47c