Use thread pool executor for AWS API requests
So far we've cached most of the AWS API listings (instances, volumes,
AMIs, snapshots, objects) but with refreshes happening synchronously.

Since some of those methods are used as part of other methods during
request handling we make them asynchronous.

Some of the EC2 instance attributes are lazily evaluated on first access
and can trigger additional API request. Due to that we also need to
cache the `listInstances` method in addition to the API methods
mentioned above.

Change-Id: I22403699ebb39f3e4dcce778efaeb09328acd932