Update patch set 9
Patch Set 9: Code-Review-2

You sort of note this yourself, but we already have actual parity with the openstack driver: neither emits api performance stats.  But the openstacksdk library itself does.

I think at this point, it's not really any business of nodepool what the underlying libs do.  I think we should remove or update the docs that suggest there is any tie-in with openstacksdk, and call that the end of it.

But if it's absolutely necessary that we emit performance stats for api calls, then I don't think hooking into client libs like this is the right approach.  It is fragile and unecessarily complex.  I do think that we could consider an approach where we used the rate limiter (which currently does have the capability to log api performance info) to emit stats.  That would be applicable to all drivers equally, and makes more sense as it reports what nodepool sees.

Patch-set: 9
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2, 66040329aa0160692f81151b5689cdab299967b4