Use ec2_client for instances and volumes
There are two modes of use for the boto3 interface library for AWS:
resource or client calls.  The resource approach returns magic objects
representing instances of each type of resource which have their own
methods and can lazy-load relationships to other objects.  The client
approach is more low-level and returns dictionary representations of
the HTTP results.

In general in the AWS driver, we try to use resources when we can
and fall back to client methods otherwise.

When we construct our own AwsInstance objects (which translates what we
get from AWS to what nodepool expects), we filled the "az" field by
getting the availability zone from the subnet.  The subnet is an object
relationship that is lazy-loaded, which means that each time we
retrieved an instance record, we issued another API call to look up
the subnet (and this information is not cached).

To correct this, we switch to getting the subnet from the placement
field of the instance (which is more sensible anyway).

To prevent future accidental lazy-loads of un-cached data, we switch
instance and volume operations from the resource style of usage to
the client.

The result is a massive increase in performance (as the subnet lazy-load
could take 0.1 seconds for each instance running, and we very frequently
list all instances).

Change-Id: I529318896fc8096bbd9dbdac60d1a29c3ac641b6