Remove openshift functional testing
OpenShift has removed their 3.11 container images from dockerhub
and while some of them are available at quay.io, not all of them
are, and it does not appear that we are able to install 3.11 any
more.

To unblock development in nodepool, remove the functional testing.

We will rely on the existing fake-based unit tests now.

Change-Id: Ifb4b02ae1058c7549c30ab915591d4b171e3540e