Use sudo when collecting podman logs in openstack-functional test
We run the podman containers as root, so when we collect the logs,
we must do that as root too, otherwise we won't see the containers.
This is why we currently collect no container logs.

Also, increase the boot timeout from 600 to 1200 seconds to deal
with what appears to be much slower boot times of the test node.

Change-Id: Ie146599f7598459c41c2d2d885b90a702bc900bb