Correct k8s RBAC apiGroups
Several of the resources that we intend to allow users to manipulate
in k8s namespace "nodes" require an apiGroup qualifier.  Most (eg
"pods") are in the default api group delineated by the empty string,
but some (eg "deployments") are in others.  This corrects those, and
also renames the nonexistent "crontabs" to "cronjobs" which is assumed
to have been a typo.

Change-Id: I11bda3dff556a66e88bf80aad55f49972a444a7f