Rollback to 1.29/stable microk8s in functional testing
We use latest/stable by default which is also based on kubernetes 1.29,
but the latest channel released yesterday and this job fails
consistently now with connection errors to k8s. Rollback to the stable
1.29 release which was last udpated over a month ago to see if that
works better for us.

We also add k8s log collection in order to debug these failures further
if necessary.

Change-Id: I72310521bdabfc3e34a9f2e87ff80f6d7c27c180