Add pod image username attribute
If a pod's container image specifies a USER layer, then we need to
tell Zuul about that, otherwise Zuul will configure ansible to
use the executor default user (which may be zuul, or may be root).
If the container user is root, then it doesn't matter since the
user will be able to write anywhere regardless, but if it is
something other than root (eg, "ubuntu"), then Ansible will try
to make tmpdirs under /root and fail.

To correct this, allow nodepool to include a username attribute
for the node so that Zuul configures ansible_user correctly for
the container image in use.

Change-Id: Ia2e8d4c75b1b78406b4512051bd7ab2720e90ca1