Add rackspaceauth as a nodepool dependency
Rackspace has announced that MFA will be required starting on March 26,
2024. When MFA is enabled on an account you will no longer be able to
log in to rackspace using a username and password with
openstacksdk/openstackclient/etc as the APIs apparently don't support
negotiating the MFA token. Instead we can either use a rackspace
specific api_key or a keystone bearer token.

We opt for the rackspace specific api_key because it doesn't expire
like the bearer tokens do. But using the rackspace api_key does require
a keystoneauth1 plugin called `rackspaceauth` to be installed which this
change adds to nodepool.

This new dep is Apache2 licensed according to the License file in the
sdist. The new dep has minimal deps of its own and they are all alread
shared by the existing dep tree. Seems reasonable to install this small
lib in hopes that we can keep rackspace working with nodepool

As a final note the OpenDev team plans to test use of the api_key with
this library against a single rackspace region. It is possible this
won't work out of the box and we may need to make additional updates.
Unfortunately, it isn't easy to test this without talking directly to
rax so we opt for the lib install and testing via OpenDev.

Change-Id: Ibff32bb44e05413391dd7a320ba356f521bb30e8