Add max-age to metastatic driver
This allows the metastatic driver to gracefully remove a backing
node from service after a certain amount of time.  This forced
retirement can be used to periodically ensure that fresh backing
nodes are used even in busy systems (which can help ensure that,
over time, job behavior does not change based on the contents of
the backing node server).

Change-Id: I62a95411a5d0b75185739a3c2553c75124c78c25