Introduce use-internal-ip parameter to OpenStack
In OpenStack releases older than Zed, there is no
interface_ip parameter that is used to get correct
IP address and nodes are failing to start with missing
IP address if auto-floating-ip is set.
To not change driver logic, use-internal-ip parameter is
introduced as WA to be able to use VMs with private IPs
on OpenStack releases older than Zed.

Change-Id: Idfcbd1681d7935a08c7a3a91e44a8b2d9139230b