Use cached node list in createMinReady
We previously avoided using the cached node list when deciding
whether to create a min-ready node request.  This is because if
the cache is out of date, we might end up issuing a min-ready
request when it isn't necessary.

However, in systems with large numbers of nodes, this can be a
very expensive operation.  Meanwhile, by introducing the tree
cache, we have attempted to make the cache as fast as possible
and therefore more likely to be up-to-date (especially in a
smaller system where this would matter more).

Let's try using the cache for this instead so that we get optimal
performance in small and large systems.

Since this is the last call site for this method which did not use
the cache, the method argument indicating whether to use the cache
is removed and unit tests adjusted accordingly.

Change-Id: I203206413265fa5216fccec0f7579e24f0aee48e