Continue trying uploads for other providers after failure
In order to avoid spending hours or days uploading images with a
stale configuration, we exit the upload loop after each successful
upload, refresh our config, and then resume where we left off.

But this assumes that uploads eventually succeed; otherwise if we
attempt an upload and it fails, we will still restart the loop and
therefore re-attempt the same upload.

To avoid this situation, only exit the loop when an upload succeeds.

Change-Id: Ib4422382a0d2abbfa55c83922b68c1b4df482dcb