Ensure nodepool-builder works with an empty secure.conf
Since I929bdb757eb9e077012b530f6f872bea96ec8bbc, when the secure.conf
file is empty, the checkConfigRecent always returns False, resulting in
the service appearing stuck as it always wait for a new config.

This change ensure that the service works, even when the secure.conf
file is empty.

Change-Id: Ic115e0b3521b123c327fb1c0f8e6238089c4c536