Support EC2 Fleet API
This change adds support for launching Amazon EC2 instances
(On-Demand and Spot), using the EC2 fleet API.

One key difference comparing to other type of labels is the quota
information can not be determined by the label itself, it is only
known after the instance is launched.

Therefore, when EC2 fleet API is configured, quota is not checked
before launching the instance, but the quota would be taken into
account after the instance is launched.

This also requires change in the function `getQuotaForLabel()` in
the `Adapter` interface and the function `quotaNeededByLabel()` in the
`QuotaSupport` interface to add an optional `instance` parameter, which
should convey the instance type information.

Change-Id: I1759d3539ef79d4a556661898553396d85aa69fa