Fix exception for non Statemachine providers
For non Statemachine providers, following exception could occur:
Traceback (most recent call last):
  File "/opt/nodepool/lib/python3.11/site-packages/nodepool/driver/utils.py", line 375, in estimatedNodepoolQuotaUsed
    if node.resources['cores']:
       ~~~~~~~~~~~~~~^^^^^^^^^
TypeError: 'NoneType' object is not subscriptable

It is because the 'node.resrouces' is only set in Statemachine
code. It is fixed by checking if 'node.resources' is None.

Mean while, we do not need to check the 'node.resources['cores']',
because for fleet api, if it is 0, means it is not lauched yet, and
it's correct to return the empty qi. For others, it is not possible
for it to be 0.

Change-Id: Ibe3795be3082cac358ca68fba6b8bdb52af3d7dc