Avoid recursive delete in forceUnlockNode
The forceUnlockNode method has exactly one call site: unlocking
the backing node within the metastatic driver.  This is because
the metastatic driver acquires a non-ephemeral lock (so that the
lock is held even if the launcher is restarted).  The Kazoo lock
recipe does not have a facility for unlocking a non-ephemeral lock
once the original lock python object is lost.  That means we need
to remove the lock znode ourselves.  That's why we have a
forceUnlockNode method.

This method previously performed a simple recursive delete on the
lock path, which would delete all of the lock contenders including
the one holding it (typically, this would be the only contender)
and also the lock path znode itself.  This is a simple, brute-force
method of handling it.

However, it is subject to a race condition where two actors might:

A] call forceUnlockNode
A] delete node/lock/contender0
B] create node/lock/contender1
A] delete node/lock

And the last call would fail because the lock znode was no longer
empty.

To address this, instead of deleting the entire lock znode, we will
delete only the node which actually has the lock.  Assuming we're
dealing with a standard exclusive lock, this is the first contender
(the node with the lowest id).  The updated method merely deletes
this lock.

This avoids the race condition above.  It also makes things less
disruptive for other actors which may be trying to acquire the lock
(in that we will no longer delete their lock contender znode).

Change-Id: I8c1c3a3fe7f14eb605b19f7ede4931c38c67393d