Allow to pass tolerations to Kubernetes pods
This will allow to specify custom tolerations to nodes with label type
pod, which will set the tolerations on the Pods allowing to indicate
which node taints are tolerated by each node type.

This is needed when a specific kubernetes nodepool is used for CI
workloads while other nodes are reserved to running system pods.

Signed-off-by: Flavio Percoco <flavio.percoco@flyrlabs.com>
Change-Id: Ieb3103506e059c0a8b439fb755f0d7c243150f79