Check ephemeral locks during request handling
The only explicit handling nodepool has for session loss is if we
encounter a session loss during a ZK operation.  If the loss happens
while the launcher is idle, Nodepool may not recover gracefully.

In particular, it may attempt to continue operating on requests or
nodes for which it has lost the lock.  And if doing so results in
an error (such as NoNodeError after another launcher cleans up
the unlocked nodes) it may loop indefinitely.

To address this, copy over the SessionAwareLock from Zuul and use
it to verify that we retain the lock before processing any request
or node.

This is not a complete solution; that would require checking the
lock before each write.  That may be better left for the nodepool-in-
zuul work.  This should address most circumstances, avoid infinite
loops, and at least limit the fallout to a single node or request.

Change-Id: I3d1878ac4579e3b9ca22f613a32399b65bda9bdf