Make AWS create_fleet more robust
Due to the AWS EC2 eventual consistency [1] it can happen that the
describe-instance directly after create-fleet fails because the instance
cannot be found. For this we should add a back-off and try again
like it is described in the AWS documentation.

This should fix failed launch attempts caused by:
botocore.exceptions.ClientError: An error occurred
(InvalidInstanceID.NotFound) when calling the DescribeInstances
operation: The instance ID 'i-0123' does not exist

or like this:
  File "/opt/nodepool/lib/python3.11/site-packages/nodepool/driver/aws/adapter.py", line 1666, in _createInstance
    return self._createFleet(label, image_id, tags, hostname, log)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/opt/nodepool/lib/python3.11/site-packages/nodepool/driver/aws/adapter.py", line 1860, in _createFleet
    instance_id = resp['Instances'][0]['InstanceIds'][0]
                  ~~~~~~~~~~~~~~~~~^^^
IndexError: list index out of range

[1] https://docs.aws.amazon.com/ec2/latest/devguide/eventual-consistency.html

Change-Id: Ic25efe72115fe82381e891115e32e9cdd188a0c2