Node instances: label modifiers (aws spot, fleet)
An instance property has been added to keep track of certain
modifiers that change the behavior of node requests. The modifiers,
delivered via config labels, are meant to track the AWS node
requests for spot instances or fleet.

Tracking these labels will provide better reporting of node
usage within Zuul jobs and the capability of jobs to react
on certain node types.

The label_mods property is expected to be accessible in Zuul
under the `nodepool` host facts.

Change-Id: I2d91f8d336738fc65ddaa26f3de6db5cba09ac0d