Set git repo ownership for nodepool dib integration testing
We run nodepool-builder as root in the integration testing but
/home/zuul/src repos are owned by zuul. When git tries to do local
clones of these repos it fails because security concerns mean it will
not talk to git repos owned by a different user.

Address this by chowning /home/zuul/src and its contents to root:root.
We don't use the git config method because that would require modifying
the container image and that seems less ideal than making the test
environment match what we need.

Change-Id: Idf78c50180b7b989082befe797f5003ebb29ec5b