Add option to specify volume_type in openstack driver
The openstack driver lets users specify boot_from_volume
together with a volume_size property to let openstack
create a volume on-the-fly.
However, the upstream openstack API does not currently allow
users to specify a volume_type in that call.

This change adds a volume_type option for the openstack
driver and, when used with boot_from_volume, creates one
with the given volume_size AND volume_type at runtime.
The volume is then passed as boot_volume which the
upstream API supports.
Other drivers do already allow the user to specify a
volume_type.

Change-Id: I2abb92d97d9fe8184e7801eb7a118b5c5b8817af
Signed-off-by: Gondermann <gondermann@b1-systems.de>